backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.16
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.8
packaging==24.1
pillow==11.0.0
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==2.2.3
django-heroku==0.3.1 
whitenoise==6.7.0
black==24.8.0
dj-database-url==2.2.0
drf-yasg==1.21.7
attrs==24.2.0 
fuzzysearch==0.7.3
django-cors-headers==4.4.0 



